Read param as in for py_ped_unit_get_name()
'i' was defined as a long, but we were reading it as an int with
PyArg_ParseTuple().  On 64-bit systems...well, you know how the story
ends.

+++ src/pyunit.c
@@ -82,7 +82,7 @@ PyObject *py_ped_unit_get_size(PyObject *s, PyObject *args) {
 
 PyObject *py_ped_unit_get_name(PyObject *s, PyObject *args) {
     const char *name;
-    long unit;
+    int unit;
 
     if (!PyArg_ParseTuple(args, "i", &unit)) {
         return NULL;
